Update pendo IIFE handling to evaluate configured IIFE
The pendo feature now evaluates the IIFE passed in via the MFE
configuration, and passes it two arguments, first is the base URL of the
lms in `lmsUrl`, and the second is the username in `username`.
Note that no error handling is provided for the IIFE, any exceptions
within the IIFE will be passed up to the browser. This assists in
debugging issues with the IIFE but risks causing errors for end users in
the MFE if the IIFE does not catch exceptions.